+++ familyhelper-life-sdk/src/main/java/com/dwarfeng/familyhelper/life/sdk/util/Constants.java
@@ -29,11 +29,30 @@ public final class Constants {
     @ComparatorItem
     public static final int COMPARATOR_DESC = 1;
 
+    /**
+     * @since 1.1.1
+     */
+    @RemindScopeTypeItem
+    public static final int REMIND_SCOPE_TYPE_OWNER_ONLY = 0;
+
+    /**
+     * @since 1.1.1
+     */
+    @RemindScopeTypeItem
+    public static final int REMIND_SCOPE_TYPE_OWNER_AND_MODIFIER = 1;
+
+    /**
+     * @since 1.1.1
+     */
+    @RemindScopeTypeItem
+    public static final int REMIND_SCOPE_TYPE_ALL_PERMITTED = 2;
+
     private static final Logger LOGGER = LoggerFactory.getLogger(Constants.class);
     private static final Lock LOCK = new ReentrantLock();
 
     private static List<Integer> permissionLevelSpace = null;
     private static List<Integer> comparatorSpace = null;
+    private static List<Integer> remindScopeTypeSpace = null;
 
     /**
      * 获取权限等级的空间。
@@ -117,6 +136,49 @@ public final class Constants {
         }
 
         comparatorSpace = Collections.unmodifiableList(result);
+    }
+
+    /**
+     * 获取提醒范围类型的空间。
+     *
+     * @return 提醒范围类型的空间。
+     * @since 1.1.1
+     */
+    public static List<Integer> remindScopeTypeSpace() {
+        if (Objects.nonNull(remindScopeTypeSpace)) {
+            return remindScopeTypeSpace;
+        }
+        // 基于线程安全的懒加载初始化结果列表。
+        LOCK.lock();
+        try {
+            if (Objects.nonNull(remindScopeTypeSpace)) {
+                return remindScopeTypeSpace;
+            }
+            initRemindScopeTypeSpace();
+            return remindScopeTypeSpace;
+        } finally {
+            LOCK.unlock();
+        }
+    }
+
+    private static void initRemindScopeTypeSpace() {
+        List<Integer> result = new ArrayList<>();
+
+        Field[] declaredFields = Constants.class.getDeclaredFields();
+        for (Field declaredField : declaredFields) {
+            if (!declaredField.isAnnotationPresent(RemindScopeTypeItem.class)) {
+                continue;
+            }
+            Integer value;
+            try {
+                value = (Integer) declaredField.get(null);
+                result.add(value);
+            } catch (Exception e) {
+                LOGGER.error("初始化异常, 请检查代码, 信息如下: ", e);
+            }
+        }
+
+        remindScopeTypeSpace = Collections.unmodifiableList(result);
     }
 
     private Constants() {